+++ source/dib.cpp
@@ -296,7 +296,7 @@ bool DIB::ApplySoftAA(DIB& target)
 		for (int x = 0; x < target.getW(); x++) {
 			pixel color[] = { getPixel(x * 2, y * 2), getPixel(x * 2 + 1, y * 2), getPixel(x * 2, y * 2 + 1), getPixel(x * 2 + 1, y * 2 + 1) };
 			ivec4 sum;
-			for (int i = 0; i < dimof(color); i++) {
+			for (int i = 0; i < (int)dimof(color); i++) {
 				sum += uint32ToIvec4(color[i]);
 			}
 			sum /= 4;
